refactor: Move range slider into it's own component
